removed displayarea and merged into layout for easier sidebar toggling
