--- src/assets/components/DisplayArea/DisplayArea.scss
@@ -1,57 +1,0 @@
-@use "../../styles/variables" as *;
-@import "../../styles/media";
-
-.display-area {
-    justify-content: center;
-    align-items: center;
-    background-color: $color-orange;
-    margin: 5% auto;
-    border-radius: 15px;
-    width: 100%;
-    color: $color-dark;
-    position: relative;
-
-    @include laptop-desktop {
-        display: flex;
-        flex-direction: column;
-        width: 80%;
-        height: 75vh;
-        margin: 1% auto;
-    }
-
-    &__loading,
-    &__error {
-        font-size: 1.4rem;
-        color: $color-dark;
-    }
-
-    &__asteroid-image {
-        width: 20%;
-        animation: spin 5s infinite linear;
-        display: block;
-        margin: 0 auto;
-    }
-
-    @keyframes spin {
-        from {
-            transform: rotate(0deg);
-        }
-        to {
-            transform: rotate(360deg);
-        }
-    }
-
-    &__asteroids {
-        flex-grow: 1;
-        width: 100%;
-        display: flex;
-        justify-content: center;
-        align-items: center;
-    }
-
-    &__nav {
-        width: 100%;
-        display: flex;
-        justify-content: center;
-    }
-}
