+++ src/assets/components/SideBar/SideBar.scss
@@ -2,28 +2,30 @@
 @import "../../styles/media";
 
 .side-bar {
-    display: none;
+    // position: absolute;
+    // z-index: 2;
+    // background-color: $color-yellow;
+    // margin-top: 5%;
+    // border-radius: 15px;
+    // width: 90%;
+    // display: flex;
+    // justify-content: center;
+    // height: 100vh;
+    // padding: 0 3rem;
+    // box-sizing: border-box;
 
-    @include laptop-desktop {
-        background-color: $color-yellow;
-        margin-top: 1%;
-        margin-right: 1%;
-        border-radius: 15px;
-        width: 20%;
-        display: flex;
-        justify-content: center;
-        height: 75vh;
-    }
-
-    &__close-icon {
-        z-index: 1001;
-        width: 25px;
-        cursor: pointer;
-
-        @include laptop-desktop {
-            display: none;
-        }
-    }
+    // @include laptop-desktop {
+    //     position:relative;
+    //     background-color: $color-yellow;
+    //     margin-top: 5%;
+    //     border-radius: 15px;
+    //     width: 100%;
+    //     display: flex;
+    //     justify-content: center;
+    //     height: 75vh;
+    //     padding: 0 3rem;
+    //     box-sizing: border-box;
+    // }
 
     &__filters {
         display: flex;
