+++ src/assets/containers/Layout/Layout.scss
@@ -17,5 +17,127 @@
     &__content {
         display: flex;
         width: 100%;
+        position: relative;
+    }
+
+    &__side-bar {
+
+         position: absolute;
+         z-index: 5;
+         margin-top: 5%;
+        top: 0;
+        left: 0;
+        width: 100%;
+      
+        background-color: $color-yellow;
+        border-radius: 15px;
+        padding: 2rem 3rem;
+        box-sizing: border-box;
+        transition: transform 0.3s ease;
+        transform: translateX(-100%); // Hidden by default
+
+        &.visible {
+            transform: translateX(0); // Slide in
+        }
+
+        &.hidden {
+            transform: translateX(-120%); // Slide out
+        }
+
+        @include tablet {
+            width: 50%;
+        }
+
+        @include laptop-desktop {
+        position:relative;
+        background-color: $color-yellow;
+        margin-top: 1%;
+        margin-right: 1%;
+        border-radius: 15px;
+        width: 20%;
+        display: flex;
+        justify-content: center;
+        height: 75vh;
+        padding: 0 3rem;
+        box-sizing: border-box;
+            transform: translateX(0); // Always visible on larger screens
+        }
+    }
+
+    &__display-area {
+        justify-content: center;
+        align-items: center;
+        background-color: $color-orange;
+        margin: 5% auto;
+        border-radius: 15px;
+        width: 100%;
+        color: $color-dark;
+        position: relative;
+
+        @include laptop-desktop {
+            display: flex;
+            flex-direction: column;
+            justify-content: flex-end;
+            width: 80%;
+            height: 75vh;
+            margin: 1% 0 1% auto;
+            padding: 0;
+        }
+    }
+
+    &__hamburger {
+        position: absolute;
+        top: 20px;
+        left: 20px;
+        cursor: pointer;
+        z-index: 10;
+
+        .layout__hamburger-bar {
+            display: block;
+            width: 30px;
+            height: 3px;
+            background-color: $color-dark;
+            margin: 5px 0;
+        }
+
+        @include laptop-desktop {
+            display: none;
+        }
+    }
+
+    &__loading,
+    &__error {
+        font-size: 1.4rem;
+        color: $color-dark;
+    }
+
+    &__asteroid-image {
+        width: 15%;
+        animation: spin 5s infinite linear;
+        display: block;
+        margin: 0 auto;
+    }
+
+    @keyframes spin {
+        from {
+            transform: rotate(0deg);
+        }
+        to {
+            transform: rotate(360deg);
+        }
+    }
+
+    &__asteroids {
+        flex-grow: 1;
+        width: 100%;
+        display: flex;
+        justify-content: center;
+        align-items: center;
+    }
+
+    &__nav {
+        width: 100%;
+        display: flex;
+        justify-content: center;
     }
 }
